converted the newsletter module to use the action controllers
