+++ ACP3/Modules/ACP3/Newsletter/Controller/Admin/Index/Index.php
@@ -1,0 +1,89 @@
+<?php
+/**
+ * Copyright (c) 2016 by the ACP3 Developers. See the LICENCE file at the top-level module directory for licencing details.
+ */
+
+namespace ACP3\Modules\ACP3\Newsletter\Controller\Admin\Index;
+
+use ACP3\Core;
+use ACP3\Modules\ACP3\Newsletter;
+
+/**
+ * Class Index
+ * @package ACP3\Modules\ACP3\Newsletter\Controller\Admin\Index
+ */
+class Index extends Core\Modules\AdminController
+{
+    /**
+     * @var \ACP3\Modules\ACP3\Newsletter\Model\NewsletterRepository
+     */
+    protected $newsletterRepository;
+
+    /**
+     * Index constructor.
+     *
+     * @param \ACP3\Core\Modules\Controller\AdminContext               $context
+     * @param \ACP3\Modules\ACP3\Newsletter\Model\NewsletterRepository $newsletterRepository
+     */
+    public function __construct(
+        Core\Modules\Controller\AdminContext $context,
+        Newsletter\Model\NewsletterRepository $newsletterRepository)
+    {
+        parent::__construct($context);
+
+        $this->newsletterRepository = $newsletterRepository;
+    }
+
+    /**
+     * @return array
+     */
+    public function execute()
+    {
+        $newsletter = $this->newsletterRepository->getAllInAcp();
+
+        /** @var Core\Helpers\DataGrid $dataGrid */
+        $dataGrid = $this->get('core.helpers.data_grid');
+        $dataGrid
+            ->setResults($newsletter)
+            ->setRecordsPerPage($this->user->getEntriesPerPage())
+            ->setIdentifier('#newsletter-data-grid')
+            ->setResourcePathEdit('admin/newsletter/index/edit')
+            ->setResourcePathDelete('admin/newsletter/index/delete');
+
+        $dataGrid
+            ->addColumn([
+                'label' => $this->translator->t('system', 'date'),
+                'type' => Core\Helpers\DataGrid\ColumnRenderer\DateColumnRenderer::class,
+                'fields' => ['date'],
+                'default_sort' => true
+            ], 50)
+            ->addColumn([
+                'label' => $this->translator->t('newsletter', 'subject'),
+                'type' => Core\Helpers\DataGrid\ColumnRenderer\TextColumnRenderer::class,
+                'fields' => ['title'],
+            ], 40)
+            ->addColumn([
+                'label' => $this->translator->t('newsletter', 'status'),
+                'type' => Core\Helpers\DataGrid\ColumnRenderer\ReplaceValueColumnRenderer::class,
+                'fields' => ['status'],
+                'custom' => [
+                    'search' => [0, 1],
+                    'replace' => [
+                        $this->translator->t('newsletter', 'not_yet_sent'),
+                        $this->translator->t('newsletter', 'already_sent'),
+                    ]
+                ]
+            ], 30)
+            ->addColumn([
+                'label' => $this->translator->t('system', 'id'),
+                'type' => Core\Helpers\DataGrid\ColumnRenderer\IntegerColumnRenderer::class,
+                'fields' => ['id'],
+                'primary' => true
+            ], 10);
+
+        return [
+            'grid' => $dataGrid->render(),
+            'show_mass_delete_button' => count($newsletter) > 0
+        ];
+    }
+}
